Ticket #16107 - make domain_id to be BIGINT

+++ sql/notify_new.sql
@@ -77,7 +77,7 @@ comment on column notify_statechange.mail_id is 'email with result of notificati
 CREATE TABLE notify_outzone_unguarded_domain_additional_email (
   crdate TIMESTAMP WITHOUT TIME ZONE NOT NULL DEFAULT now(),
   state_id BIGINT REFERENCES object_state (id),
-  domain_id INTEGER NOT NULL REFERENCES object_registry (id),
+  domain_id BIGINT NOT NULL REFERENCES object_registry (id),
   email varchar(1024) NOT NULL,
   CONSTRAINT notify_outzone_unguarded_domain_additional_email_unique_key UNIQUE (state_id, domain_id, email)
 );
